self_managing_systems/microservice/paper_artifact_arXiv_2407_14402/requirements.txt
colorlog==6.8.2
grafanalib==0.7.1
Jinja2==3.1.4
jsonpath_ng==1.6.1
msal==1.28.0
openai==1.44.1
pika==1.3.2
prometheus_api_client==0.5.5
promptflow==1.16.1
promptflow_tools==1.4.0
pyautogen==0.3.1
PyYAML==6.0.1
termcolor==2.4.0
timeout_decorator==0.5.0
httpx==0.27.2